Refactored react file structure to be nested folders by column
