+++ src/components/leftColumn/CollectionsArea/CollectionManager.tsx
