+++ src/components/middleColumns/chatbot/config.tsx
