+++ src/components/middleColumns/middleTopCol/StudyComponent.tsx
