+++ src/components/rightColumn/FilePreview.tsx
