+++ src/components/rightColumn/FilePreviewer.tsx
