+++ src/components/rightColumn/RightColumn.tsx
