+++ src/components/rightColumn/SummaryViewer.tsx
